requirements.txt
numpy>=1.19.0
scipy>=1.5.0
python-dateutil>=2.8.0
